requirements.txt
azure-identity>=1.12.0
azure-mgmt-costmanagement>=1.0.0
azure-mgmt-resource>=23.0.1
pandas>=2.0.0
openpyxl>=3.1.2
python-dotenv>=1.0.0 